Checkboxes working, mapped over disc type to make checkbox for each option
